pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agents"
version = "0.1.0"
description = "SDLC agent system for GitHub: Code Agent and AI Reviewer Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "PyGithub>=2.0.0",
    "GitPython>=3.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=24.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
]

[project.scripts]
gaj = "coding_agents.cli:main"
code-agent = "coding_agents.cli_code_agent:main"
reviewer-agent = "coding_agents.cli_reviewer:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
